Regression model Sales Forecasting
i. You are provided with historical sales data for 45 retail stores located in 
different regions. Each store contains a number of departments, and you are 
tasked with predicting the department-wide sales for each store.
ii. In addition, the retail company runs several promotional markdown events 
throughout the year. These markdowns precede prominent holidays, the four 
largest of which are the Super Bowl, Labour Day, Thanksgiving, and 
Christmas. 
iii. Data:
1. stores.csv: This file contains anonymized information about the 45 
stores, indicating the type and size of the store. 
2. features.csv: Comprise of information on Markdown and 
macroeconomic factors.
3. train.csv: This is the historical training data.
iv. Write the performance and key insights on the data in a word doc